generator: support creating a unified EFI file
This provides the scaffolding required in order to support Secure Boot in the
installer.

Specifically, we now make a distinction between a `BootableToplevel` and an
`EfiProgram`, which is just a thin wrapper around a `BootableToplevel` with
functionality to combine the /etc/os-release, kernel cmdline, kernel proper, and
initrd into a single EFI bootable using the systemd-boot EFI stub. This way, we
only need to sign one file, and booting via that EFI image will guarantee the
kernel cmdline, kernel proper, and initrd have not been tampered*.

One downside to combining all this information into a singular EFI image is that
there is no way to "deduplicate" them, seeing that the same combination of
kernel and initrd might have a different cmdline (thus, we name the EFI image
after the Nix store hash of the toplevel). The end result is that *every
generation* will now take up `os-release_size + cmdline_size + kernel_size +
initrd_size` in the ESP.

* Of course, this relies on the signing keys being properly guarded, etc.