installer: patch sbattach to allow padding unsigned files
This is required for the moment so that we can both verify the file
signatures as well as the hash of the underlying file. sbsigntool pads
the EFI files for some compatibility reason, but removing the signature
with `sbattach --remove` does not remove this padding either (likely
because there's no way to tell if the padding was added by the
sbsigntool suite or the user or something else).