installer: support secure boot signed files
In order to support Secure Boot, this implementation adds 2 new steps:
`ReplaceFiles` and `SignFiles`.

The `ReplaceFiles` step determines which files need to be actually
replaced:

    1. If the files are signed for Secure Boot, they are verified with
    `sbverify`.

        a. If they are successfully verified, they are then copied to a
        temporary location and stripped of their signatures in order to
        verify they are the same file (using a CRC32 checksum).

        b. If they are the same file according to the checksums, the
        file is removed from the `generated_entries` directory so that
        it won't overwrite the same file in the ESP (or waste space).

        c. If they are not the same file according to the checksums, the
        file will be overwritten by the `generated_entries` version in
        the `CopyToEsp` step.

    2. If the files are not signed for Secure Boot, they undergo a
    simple hash check (using a CRC32 checksum).

        a. If they are the same file according to the checksums, the
        file is removed from the `generated_entries` directory so that
        it won't overwrite the same file in the ESP (or waste space).

        b. If they are not the same file according to the checksums, the
        file will be overwritten by the `generated_entries` version in
        the `CopyToEsp` step.

The `SignFiles` step is given a list of EFI files to sign and the
`SigningInfo` struct and does as its name implies: using the provided
`sbsign` and key and certificate, it signs all the existing EFI files
for use with Secure Boot.

Because the `CopyToEsp` step copies the `generated_entries` directory to
the ESP (first, to a temporary staging directory, followed by atomically
moving each file into place), the `ReplaceFiles` step really only needs
to take care of (read: remove) the files that are *not* going to be
replaced.