systemd_boot/plan: sign files before pruning them
Due to the generator and the installer being separate, there is no
(easy) way to tell which generated files belong to which generation,
so the installer recomputes the generations by taking the latest
generations from the system profile. This means that the list of files
to sign is generated from the latest system generations, and if the
generated entries don't include some of the latest system generations
(within the limit provided by the `--configuration-limit` flag), the
list of files to sign will contain a generation that will, at that
point, have been pruned.

The easy solution is just to sign these files before we prune
everything.